Move partitioned data access points to a single file
