+++ packages/op-coreutils/src/op_coreutils/partitioned/__init__.py
@@ -2,7 +2,7 @@
 
 from .reader import DataReader, construct_input_batches
 from .location import DataLocation, MarkersLocation
-from .marker import Marker, markers_for_dates
+from .marker import Marker
 from .output import OutputData, ExpectedOutput, OutputPartMeta, KeyValue
 from .paths import get_dt, get_root_path
 from .types import SinkMarkerPath, SinkOutputRootPath
