+++ tests/op_coreutils/partitioned/test_writehelper.py
@@ -8,6 +8,7 @@ from op_coreutils.duckdb_local import run_query
 from op_coreutils.partitioned.location import DataLocation
 from op_coreutils.partitioned.writehelper import ParqueWriteManager
 from op_coreutils.partitioned.output import ExpectedOutput, OutputData
+from op_coreutils.partitioned.types import SinkOutputRootPath, SinkMarkerPath
 
 
 def test_parquet_writer():
@@ -43,9 +44,11 @@ def test_parquet_writer():
         location=DataLocation.LOCAL,
         expected_output=ExpectedOutput(
             dataset_name="daily_address_summary/daily_address_summary_v1",
-            root_path="intermediate/daily_address_summary/daily_address_summary_v1",
+            root_path=SinkOutputRootPath(
+                "intermediate/daily_address_summary/daily_address_summary_v1"
+            ),
             file_name="out.parquet",
-            marker_path="BLAH",
+            marker_path=SinkMarkerPath("BLAH"),
             process_name="default",
             additional_columns={"model_name": "MYMODEL"},
             additional_columns_schema=[
